refactor(common/release): move to controller/publisher
