+++ src/dashboard/apigateway/.importlinter
@@ -33,5 +33,3 @@ ignore_imports =
     apigateway.biz.esb.permissions -> apigateway.apps.esb.utils
     apigateway.controller.crds.release_data.release_data -> apigateway.biz.gateway_jwt
     apigateway.common.audit.shortcuts -> apigateway.apps.audit.signals
-    apigateway.common.release.publish -> apigateway.controller.tasks
-    apigateway.common.release.publish -> apigateway.controller.constants
