+++ src/dashboard/apigateway/apigateway/biz/stage.py
@@ -23,7 +23,7 @@ from django.db import transaction
 from django.utils.translation import gettext as _
 from rest_framework import serializers
 
-from apigateway.common.release.publish import trigger_gateway_publish
+from apigateway.controller.publisher.publish import trigger_gateway_publish
 from apigateway.core.constants import DEFAULT_BACKEND_NAME, DEFAULT_STAGE_NAME, PublishSourceEnum, StageStatusEnum
 from apigateway.core.models import Backend, BackendConfig, MicroGateway, Release, Stage
 from apigateway.utils.time import now_datetime
